backend/requirements.txt
# Core
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.5
pydantic-settings==2.7.1

# Database
sqlalchemy[asyncio]==2.0.37
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.14.1
pgvector==0.3.6

# Task queue
celery[redis]==5.4.0
redis==5.2.1

# PDF processing
PyMuPDF==1.25.3

# LLM / Embedding
anthropic==0.43.0
openai==1.61.0
google-genai>=1.0.0

# HTTP client (used by LLM services)
httpx==0.28.1

# SSE
sse-starlette==2.2.1

# Utilities
python-multipart==0.0.18
numpy>=1.26.0
